% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\docType{data}
\name{fasta}
\alias{fasta1}
\title{Describes Aspects of Protein Sequences}
\format{
An object of class \code{character} of length 1.
}
\usage{
sequence1
}
\arguments{
\item{AA}{(Amino Acid Sequence)}
}
\value{
Number of Amino Acids in the Sequeuce
}
\description{
Describes Aspects of Protein Sequences
}
\examples{
CountAA(sequence1)
}
\keyword{datasets}
